+++ balances/src/lib.rs
@@ -37,7 +37,6 @@ use frame_support::{
 use frame_system::{self as frame_system, ensure_signed};
 use log::{debug, info};
 use pallet_asset_tx_payment::HandleCredit;
-use pallet_transaction_payment::OnChargeTransaction;
 use sp_runtime::{traits::StaticLookup, AccountId32};
 use sp_std::convert::TryInto;
 
@@ -303,56 +302,6 @@ impl<T: Config> Pallet<T> {
 
 	pub fn purge_balances(cid: CommunityIdentifier) {
 		<Balance<T>>::remove_prefix(cid, None);
-	}
-}
-
-pub type OnChargeTransactionOf<T> = <T as pallet_transaction_payment::Config>::OnChargeTransaction;
-// Balance type alias.
-pub type BalanceOf<T> = <OnChargeTransactionOf<T> as OnChargeTransaction<T>>::Balance;
-
-pub type AssetBalanceOf<T> =
-	<<T as pallet_asset_tx_payment::Config>::Fungibles as fungibles::Inspect<
-		<T as frame_system::Config>::AccountId,
-	>>::Balance;
-pub type AssetIdOf<T> = <<T as pallet_asset_tx_payment::Config>::Fungibles as fungibles::Inspect<
-	<T as frame_system::Config>::AccountId,
->>::AssetId;
-
-pub struct BalanceToCommunityBalance<T>(PhantomData<T>);
-impl<T> BalanceConversion<BalanceOf<T>, AssetIdOf<T>, AssetBalanceOf<T>>
-	for BalanceToCommunityBalance<T>
-where
-	T: Config + pallet_asset_tx_payment::Config,
-	encointer_primitives::communities::CommunityIdentifier: From<AssetIdOf<T>>,
-	AssetBalanceOf<T>: From<u128>,
-	<T as pallet_asset_tx_payment::Config>::Fungibles:
-		fungibles::InspectMetadata<<T as frame_system::Config>::AccountId>,
-	u128: From<BalanceOf<T>>,
-{
-	type Error = Error<T>;
-
-	fn to_asset_balance(
-		balance: BalanceOf<T>,
-		asset_id: AssetIdOf<T>,
-	) -> Result<AssetBalanceOf<T>, Self::Error> {
-		let decimals =
-			<<T as pallet_asset_tx_payment::Config>::Fungibles as fungibles::InspectMetadata<
-				<T as frame_system::Config>::AccountId,
-			>>::decimals(&asset_id.into());
-
-		let fee_conversion_factor = Pallet::<T>::fee_conversion_factor();
-		let balance_u128: u128 = balance.into();
-		// 5.233 micro ksm correspond to 0.01 units of the community currency assuming a feeConversionFactor of 10_000
-		// the KSM balance parameter comes with 12 decimals
-		// 5.233 * 10^6 pKSM = 0.01 * 10^decimals LEU
-		// 5.233 * 10^6 pKSM = 0.01 * 10^(decimals - 4) * feeConversionFactor LEU
-		// 1 pKSM = (0.01 * 10^(decimals - 4) * feeConversionFactor) / (5.233 * 10^6) LEU
-		// 1 pKSM = (0.01 * 10^(decimals - 10) * feeConversionFactor) / 5.233 LEU
-		return Ok((balance_u128 *
-			(((0.01f64 / 5.233f64) *
-				10i128.pow((decimals - 10) as u32) as f64 *
-				fee_conversion_factor as f64) as u128))
-			.into())
 	}
 }
